changed where colorpalette was on design and changes props to make it work like before
