+++ components/button/continueButton/ContinueButton.js
@@ -1,9 +1,9 @@
-import Image from "next/image";
+
 import styles from "./ContinueButton.module.css";
-import {useState} from 'react';
+
 
 export const ContinueButton = ({children, onClick}) => {
-    const [activeTab, setActiveTab] = useState(0);
+   
 
     return (
         <>
